Consolidate field-scripts into bin directory.
